split registration form into pages, verify that form successfully creates experts
