+++ client/src/components/RegBasics.js
@@ -1,55 +1,38 @@
 import React, { useState } from "react";
 import { Form, Button, Col } from "react-bootstrap";
 import { CountryDropdown, RegionDropdown } from "react-country-region-selector";
-import CompanyInput from "./CompanyInput";
-import CourseInput from "./CourseInput";
 import { customFetch } from "./Helpers";
-import SkillInput from "./SkillInput";
+import { useHistory } from "react-router-dom";
 
 /* TODO: input field validations, pagify registration form */
 /* form allowing for the registration of new experts */
-export default function RegForm() {
+export default function RegBasics() {
+  let history = useHistory();
   const [staticState, setStaticState] = useState({
     firstName: "",
     lastName: "",
-    mainEmail: "",
-    tagLine: "",
+    email: "",
+    description: "",
     photoUrl: "",
     city: "",
     phone: "",
     workEmail: "",
     schoolEmail: "",
-    gitHub: "",
-    linkedIn: "",
+    gitHubUser: "",
+    linkedInUrl: "",
   });
   const [country, setCountry] = useState("");
-  const [region, setRegion] = useState("");
-  const [skills, setSkills] = useState([
-    { name: "", description: "", years: "" },
-  ]);
-  const [courses, setCourses] = useState([
-    { name: "", description: "", semester: "", year: "", grade: "" },
-  ]);
-  const [companies, setCompanies] = useState([
-    {
-      name: "",
-      description: "",
-      industry: "",
-      current: "",
-      position: "",
-      years: "",
-    },
-  ]);
+  const [state, setState] = useState("");
 
-  function handleSubmit() {
+  async function handleSubmit() {
     if (
       staticState.firstName === "" ||
       staticState.lastName === "" ||
-      staticState.mainEmail === "" ||
+      staticState.email === "" ||
       staticState.city === "" ||
       staticState.phone === "" ||
       country === "" ||
-      region === ""
+      state === ""
     ) {
       alert("required fields missing!");
       return;
@@ -58,44 +41,33 @@ export default function RegForm() {
       phone: staticState.phone,
       workEmail: staticState.workEmail,
       schoolEmail: staticState.schoolEmail,
-      gitHubUser: staticState.gitHub,
-      linkedInUrl: staticState.linkedIn,
+      gitHubUser: staticState.gitHubUser,
+      linkedInUrl: staticState.linkedInUrl,
       city: staticState.city,
-      state: region,
+      state: state,
       country: country,
     };
     let payload = {
       firstName: staticState.firstName,
       lastName: staticState.lastName,
-      email: staticState.mainEmail,
-      description: staticState.tagLine,
+      email: staticState.email,
+      description: staticState.description,
       photoUrl: staticState.photoUrl,
+      contactDetails: contactDetails,
       expertId: 0,
       active: true,
     };
-    payload = { ...payload, contactDetails, skills, courses, companies };
-    payload.skills.forEach((element, index) => {
-      if (element.name === "") {
-        payload.skills.splice(index, 1);
-      }
-    });
-    payload.courses.forEach((element, index) => {
-      if (element.name === "") {
-        payload.courses.splice(index, 1);
-      }
-      let term = element.semester + " " + element.year;
-      delete element.semester;
-      delete element.year;
-      element.term = term;
-    });
-    payload.companies.forEach((element, index) => {
-      if (element.name === "") {
-        payload.companies.splice(index, 1);
-      }
-    });
     console.log("the payload is: ", JSON.stringify(payload));
-    // customFetch("localhost:6997", "POST", payload);
-    /* TODO: redirect to profile page */
+    let response = await customFetch(
+      "http://localhost:6997/experts",
+      "POST",
+      payload
+    );
+    /* TODO: redirect to skills entry page */
+    history.push({
+      pathname: "/register2",
+      state: { expertId: response.expertId },
+    });
   }
 
   /* general change handler for all of the non-dynamic form fields */
@@ -105,84 +77,6 @@ export default function RegForm() {
       ...staticState,
       [event.target.name]: value,
     });
-  }
-
-  /* TODO: combine all of the dynamic handleChange and handleDelete functions */
-
-  /* change handler for dynamic skill inputs */
-  function handleSkillChange(event) {
-    const updatedSkills = [...skills];
-    updatedSkills[event.target.dataset.idx][
-      event.target.className.split(" ")[0]
-    ] = event.target.value;
-    setSkills(updatedSkills);
-  }
-
-  /* handler enabling the deletion of skill lines */
-  function handleSkillDelete(idx) {
-    const updatedSkills = [...skills];
-    updatedSkills.splice(idx, 1);
-    setSkills(updatedSkills);
-  }
-
-  /* creates a new empty skill array in the skills state */
-  function addSkill() {
-    setSkills([...skills, { name: "", description: "", years: "" }]);
-  }
-
-  /* change handler for dynamic course inputs */
-  function handleCourseChange(event) {
-    const updatedCourses = [...courses];
-    updatedCourses[event.target.dataset.idx][
-      event.target.className.split(" ")[0]
-    ] = event.target.value;
-    setCourses(updatedCourses);
-  }
-
-  /* handler enabling the deletion of course lines */
-  function handleCourseDelete(idx) {
-    const updatedCourses = [...courses];
-    updatedCourses.splice(idx, 1);
-    setCourses(updatedCourses);
-  }
-
-  /* creates a new empty course array in the courses state */
-  function addCourse() {
-    setCourses([
-      ...courses,
-      { name: "", description: "", semester: "", year: "", grade: "" },
-    ]);
-  }
-
-  /* change handler for dynamic company inputs */
-  function handleCompanyChange(event) {
-    const updatedCompanies = [...companies];
-    updatedCompanies[event.target.dataset.idx][
-      event.target.className.split(" ")[0]
-    ] = event.target.value;
-    setCompanies(updatedCompanies);
-  }
-
-  /* handler enabling the deletion of company lines */
-  function handleCompanyDelete(idx) {
-    const updatedCompanies = [...companies];
-    updatedCompanies.splice(idx, 1);
-    setCompanies(updatedCompanies);
-  }
-
-  /* creates a new empty company array in the courses state */
-  function addCompany() {
-    setCompanies([
-      ...companies,
-      {
-        name: "",
-        description: "",
-        industry: "",
-        current: "",
-        position: "",
-        years: "",
-      },
-    ]);
   }
 
   return (
@@ -215,8 +109,8 @@ export default function RegForm() {
             <Form.Control
               type="email"
               placeholder="Enter email"
-              name="mainEmail"
-              value={staticState.mainEmail}
+              name="email"
+              value={staticState.email}
               onChange={handleChangeStatic}
             />
           </Form.Group>
@@ -228,8 +122,8 @@ export default function RegForm() {
               as="textarea"
               rows={2}
               placeholder="FPGA Tour Winner"
-              name="tagLine"
-              value={staticState.tagLine}
+              name="description"
+              value={staticState.description}
               onChange={handleChangeStatic}
             />
             <Form.Text className="text-muted">
@@ -257,8 +151,8 @@ export default function RegForm() {
           <RegionDropdown
             disableWhenEmpty={true}
             country={country}
-            value={region}
-            onChange={setRegion}
+            value={state}
+            onChange={setState}
           />
         </Form.Row>
         <Form.Row>
@@ -311,8 +205,8 @@ export default function RegForm() {
             <Form.Label>GitHub Username</Form.Label>
             <Form.Control
               placeholder="audiophile"
-              name="gitHub"
-              value={staticState.gitHub}
+              name="gitHubUser"
+              value={staticState.gitHubUser}
               onChange={handleChangeStatic}
             />
           </Form.Group>
@@ -320,48 +214,14 @@ export default function RegForm() {
             <Form.Label>LinkedIn Profile</Form.Label>
             <Form.Control
               placeholder="https://www.linkedin.com/in/charlotte_russo"
-              name="linkedIn"
-              value={staticState.linkedIn}
+              name="linkedInUrl"
+              value={staticState.linkedInUrl}
               onChange={handleChangeStatic}
             />
           </Form.Group>
         </Form.Row>
-        <h3>Skills</h3>
-        {skills.map((val, idx) => (
-          <SkillInput
-            key={`skill-${idx}`}
-            idx={idx}
-            skillState={skills}
-            handleChange={handleSkillChange}
-            handleDelete={handleSkillDelete}
-            addSkill={addSkill}
-          />
-        ))}
-
-        <h3>Courses</h3>
-        {courses.map((val, idx) => (
-          <CourseInput
-            key={`skill-${idx}`}
-            idx={idx}
-            courseState={courses}
-            handleChange={handleCourseChange}
-            handleDelete={handleCourseDelete}
-            addCourse={addCourse}
-          />
-        ))}
-        <h3>Companies</h3>
-        {companies.map((val, idx) => (
-          <CompanyInput
-            key={`skill-${idx}`}
-            idx={idx}
-            companyState={companies}
-            handleChange={handleCompanyChange}
-            handleDelete={handleCompanyDelete}
-            addCompany={addCompany}
-          />
-        ))}
         <Button variant="primary" onClick={handleSubmit}>
-          Submit
+          Next
         </Button>
       </Form>
     </div>
